+++ src/main/java/oreo/command/ByeCommand.java
@@ -1,3 +1,8 @@
+package oreo.command;
+
+import oreo.task.TaskList;
+import oreo.ui.Ui;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(Ui ui, TaskList tasks) {
