+++ src/main/java/oreo/command/Command.java
@@ -1,3 +1,8 @@
+package oreo.command;
+
+import oreo.task.TaskList;
+import oreo.ui.Ui;
+
 public abstract class Command {
     public abstract void execute(Ui ui, TaskList tasks);
