+++ src/main/java/oreo/command/DeleteCommand.java
@@ -1,3 +1,8 @@
+package oreo.command;
+
+import oreo.exception.IllegalCommandException;
+import oreo.task.TaskList;
+import oreo.ui.Ui;
 import java.util.Scanner;
 
 public class DeleteCommand extends Command {
