+++ src/main/java/oreo/command/MarkUnmarkCommand.java
@@ -1,3 +1,9 @@
+package oreo.command;
+
+import oreo.exception.IllegalCommandException;
+import oreo.task.TaskList;
+import oreo.ui.Ui;
+
 import java.util.Scanner;
 
 public class MarkUnmarkCommand extends Command {
