+++ src/main/java/oreo/datetime/TimeParser.java
@@ -1,9 +1,11 @@
-import java.sql.Date;
+package oreo.datetime;
+
+import oreo.exception.IllegalDateTimeException;
+
 import java.time.DateTimeException;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Scanner;
 
 public class TimeParser {
     public static String parseDateOut(String input) throws IllegalDateTimeException {
