+++ src/main/java/oreo/task/Event.java
@@ -1,4 +1,7 @@
-import java.sql.Time;
+package oreo.task;
+
+import oreo.datetime.TimeParser;
+import oreo.exception.IllegalDateTimeException;
 
 public class Event extends Task {
     private String fromDate;
